requirements.txt
# Core dependencies
torch>=2.0.0
tensorflow>=2.13.0
transformers>=4.35.0
openai>=1.0.0
anthropic>=0.8.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
datasets>=2.14.0
jsonlines>=3.1.0

# API and web
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
pydantic>=2.4.0
pydantic-settings>=2.0.0

# Evaluation and metrics
scikit-learn>=1.3.0
scipy>=1.11.0
nltk>=3.8.0
langdetect>=1.0.9

# Visualization and reporting
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
jinja2>=3.1.0

# Utilities
pyyaml>=6.0.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.6.0
tqdm>=4.66.0

# Logging and monitoring
loguru>=0.7.0
sentry-sdk>=1.35.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development
black>=23.10.0
flake8>=6.1.0
mypy>=1.6.0
pre-commit>=3.5.0

# Optional GPU support
# torch-cuda>=2.0.0  # Uncomment if using CUDA